Revert "JAMES-2404 Use programmatically registered extensions"
This reverts commit f8faf50e8cb76cdd1931ddcebb0f3e4865028eaf.

This commit introduced a move toward JUNIT 5.2.0 in order to use programmatically registered extensions, which looked very promissing.
However, JUNIT 5.2.0 is not stable yet, and that lead to many tests being skept. As such, we failed finding a satisfying solution, so we have to revert, in order to get all tests played.