JAMES-2530 Rule serialization for the JMAP layer
This is a copy of the event sourcing serialization.

Please note that this was done to allow JSON format to diverge between storage and API.

For instance let's say we want to rename a field on the JMAP layer, but we would want to keep the storage untouched for limiting our work.

This DTO will be tested later in JMAP CRUD tests