EventBusContract should use verify no interaction in time window
RabbitMQ need some delay to dispatch an event to a mailbox listener.
With current `verifyZeroInteraction()`, mockito verifies immediately after
dispatch() calls. It could be alright with MemoryEventBus but not with
RabbitMQEventBus because `verifyZeroInteraction()` be okay at the moment
but not sure in latter second when events arrived to the receivers.