JAMES-2685 Specify host for file URL in LocalFileBlobExport
A fully file url has a pattern: file://<host>/<path>
It's okay to go with `file://var/blobExporting` with tests in James.
But for a real deployment, `file://var/blobExporting` is actually pointing
to `file:///blobExporting` (`var` is treated as the host, but the host is
unknown, then it omits the host and uses localhost instead).

I find it may be awkward when open the mail, and click to the url
`file://var/blobExporting` but the browser navigates me to `file:///blobExporting`