JAMES-2865 Behaviors refactoring to have chains of behaviors
Currently we just support either execute MockBehavior or build message
in the MessageHandler. But with the new requirement in case of rejecting,
if Behavior is not rejected, so as the contract describe, it can both
execute MockBehavior(pretending an error but retry success) and build
the message at the same time (because it success).

And another Behavior is compulsory if the message is matched: updating
behavior repository.

We need kind of chain of behaviors:
 - mock behavior(rejected), updating repository
 - mock behavior(not rejected), updating repository, build the mail
 - only build the mail when non matching

I decided to apply the chain of responsibility pattern and deliver a State
object across behaviors. A behavior might need to check the result of
the previous behavior in order to decide whether it should ignore/handle/
setError...

The current Behavior doesn't have the capability of:
 - having multiple behaviors executing sequentially
 - behavior doesn't have enough of information to handle because we just
 pass only input data to the behavior, nothing else. So we end up with
 constructing a behavior with some required components, it works, but ugly