[DONT REVIEW] JAMES-3012 PART-1 #2998
More precise - integration tests. The will be an attempt to move some
test class in jmap-integration-tests to where they should belong to -
webadmin-integration-tests.

To make it become smooth, I need to extract some common classes to a
generic maven module which is james-server-testing.

This is the first step, afterward is moving, and the last is cleaning