[POC] JamesServerExtension delegates params resolving to a new component
params resolving logic will grow up in the incoming future with the
capability of providing Guice Probes... as test parameters - which
requires a running James server to create these instances.

The current Extension suit is supposed to be created before James
creation. I find it's hard to adapt the about feature by writing a
new Extension. Instead, proposing a new builder method in JamesServerBuilder

The code will be looking like this:

    @RegisterExtension
    static JamesServerExtension jamesServerExtension = new JamesServerBuilder()
        .extension(new AbcExtension())
        .server(configuration -> GuiceJamesServer.forConfiguration(configuration)
            .combineWith(MemoryJamesServerMain.IN_MEMORY_SERVER_AGGREGATE_MODULE)
            .overrideWith(new WebadminIntegrationTestModule()))
        .supportsParam(DataProble.class, jamesServer -> jamesServer.getDataProbe())
        .build();