JAMES-3088 Migrating UserRepository tests to Junit 5 contract style
This is a `have to` step in order to test every method with and without
virtualHosting. It requires to have separated contracts implementing
the common one. The different is only about virtualHosting support.

Otherwise, with the incoming number of tests regarding this issue.
The test code will be really messy. And we cannot easy to test the LDAP
version because it doesn't implement the abstract test right now

Memory version go first